+++ .changeset/witty-balloons-love.md
@@ -1,0 +1,5 @@
+---
+'@undataforum/components': minor
+---
+
+Rename YouTube to Video
