+++ packages/components/src/components/Video/index.js
@@ -1,0 +1,1 @@
+export { default } from './Video';
