--- packages/components/src/components/YouTube/index.js
@@ -1,1 +1,0 @@
-export { default } from './YouTube';
