+++ packages/components/src/index.js
@@ -26,6 +26,6 @@ export {
   profileType,
   SmartLink,
   SocialIcons,
-  YouTube,
+  Video,
 } from './components';
 export { default as theme } from './theme';
